requirements.txt
Django==4.0.4
gunicorn==20.1.0
whitenoise==6.2.0
djangorestframework==3.14.0
django-cors-headers==3.13.0
deepgram-sdk==2.8.0
dj-database-url==2.0.0
channels==4.0.0
requests==2.31.0
uvicorn==0.18.3
channels-redis==4.1.0